refactors input of 09-SHA3 10-Perlin 16-PoW 19-Prime 21-PID

+++ 21-PID/p5_21_PID/PID.pde
@@ -12,7 +12,7 @@ public class PID {
   private float derivative = 0.0;
   private float input;
 
-  public PID(byte[] b) {
+  public PID(int[] b) {
     goal = 1.0 * b[b.length - 1];
     dt = 1.0 / b.length;
